+++ source/Examples/Resources/CMServiceConnectionPoint/CMServiceConnectionPoint_Absent.ps1
@@ -1,0 +1,18 @@
+<#
+    .SYNOPSIS
+        A DSC configuration script to remove a service connection point from Configuration Manager.
+#>
+Configuration Example
+{
+    Import-DscResource -ModuleName ConfigMgrCBDsc
+
+    Node localhost
+    {
+        CMServiceConnectionPoint ExampleSettings
+        {
+            SiteCode       = 'Lab'
+            SiteServerName = 'CA01.contoso.com'
+            Ensure         = 'Absent'
+        }
+    }
+}