+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added CMDsitributionPoint Resource
 - Added CMHeartbeatDiscovery module
 - Added ConvertTo-ScheduleInterval to ResourceHelper
+- Added CMServiceConnectionPoint Resource
 
 ### Changed